requirements-dev.txt
# 开发和测试依赖

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# AWS Mock
moto>=4.2.0

# 代码质量工具
black>=23.12.0
pylint>=3.0.0
flake8>=6.1.0

# 类型检查
mypy>=1.7.0
boto3-stubs[sqs,dynamodb,lambda]>=1.34.0

# 开发工具
ipython>=8.18.0
